+++ src/main/java/com/ecomers/api/restaurants/web/controller/CommerceController.java
@@ -67,7 +67,7 @@ public class CommerceController {
 
     //......................................................................................
     @GetMapping("/{id}/rate/{distance}")
-    public ResponseEntity<Double> getRate(@PathVariable("id") int id,
+    public ResponseEntity<Integer> getRate(@PathVariable("id") int id,
                                           @PathVariable("distance") double distance  ) {
         return service.getRate(id, distance*1000)
                 .map(rate -> new ResponseEntity<>(rate, HttpStatus.OK))
